Add API versioning to the project. Update package.json version to 1.0.9, modify route manager to use v1 routes, and move hello and user routes to v1 subdirectory. Update Swagger documentation to include v1 endpoints
