+++ route/v1/user.route.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
-const userController = require('../controller/user.controller.js');
-const validator = require('../middleware/validators.js');
-const auth = require("../middleware/auth");
+const userController = require('../../controller/user.controller.js');
+const validator = require('../../middleware/validators.js');
+const auth = require("../../middleware/auth");
 
 router.post('/signin', validator.validateSignIn, userController.signIn);
 router.post('/signup', validator.validateSignUp, userController.signUp);
